+++ src/Components/Query.php
@@ -6,7 +6,7 @@
  * @subpackage League\Uri\Components
  * @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
  * @license    https://github.com/thephpleague/uri-components/blob/master/LICENSE (MIT License)
- * @version    1.0.4
+ * @version    1.1.0
  * @link       https://github.com/thephpleague/uri-components
  *
  * For the full copyright and license information, please view the LICENSE
@@ -130,9 +130,6 @@ class Query implements ComponentInterface, Countable, IteratorAggregate
         }
 
         $str = $this->validateString($str);
-        if (false !== strpos($str, '#')) {
-            throw new Exception(sprintf('The encoded query `%s` contains invalid characters', $str));
-        }
 
         return static::parse($str, static::$separator);
     }